requirements.txt
PySide6-Essentials
pyinstaller